requirements.txt
keras==2.11.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
plotly==5.13.1
python-dotenv==1.0.0
requests==2.28.2
scikit_learn==1.2.2
stocksymbol==0.0.5
streamlit==1.20.0
streamlit_lottie==0.0.3
yfinance==0.2.14
streamlit==1.20.0
fontawesome-free